Refactor Canvas component and introduce new hooks for part operations and canvas events
- Updated import paths for usePartOperations and useCanvasEvents hooks to streamline part management and event handling.
- Added useCanvasEvents hook to manage mouse interactions and camera movements within the Canvas component, enhancing drag-and-drop functionality.
- Introduced usePartOperations hook to centralize socket operations for adding, updating, and deleting parts, improving code organization and maintainability.

These changes enhance the structure and functionality of the prototype editing interface, ensuring a more efficient user experience.
