+++ frontend/src/features/prototype/hooks/useCanvasEvents.ts
